+++ src/message_provider/relay/relay_client.py
@@ -139,14 +139,16 @@ class RelayClient:
                 # Hub wants to add a reaction through local provider
                 self.local_provider.send_reaction(
                     message_id=data.get('message_id'),
-                    reaction=data.get('reaction')
+                    reaction=data.get('reaction'),
+                    channel=data.get('channel')
                 )
 
             elif msg_type == 'update_message':
                 # Hub wants to update a message through local provider
                 self.local_provider.update_message(
                     message_id=data.get('message_id'),
-                    new_text=data.get('new_text')
+                    new_text=data.get('new_text'),
+                    channel=data.get('channel')
                 )
 
             else: